rename NeuronTag component
